Set system locale to international
Running `free` without any modifications will result in the output being
localized into the current system-language (i.e. German if you're running a
German install of Linux). The parsing of the `memory` script however relies
on the output being in english (US).

As an easy fix we just set the system locale tempoararily to 'C' (which is the
_default_ language, i.e. english in case of `free`).